[signature] don't fail on unknown requirements
Mach-O files with unsupported requirement types ( e.g. Library )
fail to parse as types.ParseRequirements returns an error.

Currently, unknown opcodes generates an error message in
the parsed requirement detail.
This commit makes the behavious similar also at the higher level:
requirement type.

In this way, we end up with an error message in the parsed requirement,
but we'll still be able to parse and use the rest of the file.