[symtab] do not change read symbols
The initial codebase of this project was taken from golang
compiler / builder, which assumed that read / written Mach-Os
are created by/for go applications.

Due to an issue with the linker, the go compiler adds a "_" at
the beginning of any golang symbol.
When it reads it back, it undoes this addition.

I've removed this logic as the purpose of this library is to parse
generic Mach-O files to inspect their content, regardless of the
application you use them against.

ref: https://github.com/golang/go/issues/33808
ref: https://go-review.googlesource.com/c/go/+/196217